Disable post format input if the theme does not support post formats.
This is the same behaviour we have currently on the classic. Showing the post format input also has another non-desirable side effect, when the theme does not support post formats, posts are created with standard post format and not the default post format setting. As the setting is not respected, leaving the input there confuses the user.